Add command to restart the 7Days2Die server
Accesses a server instance and restarts it. Has some very strict requirements to work specifically for my own server. Requires the private key file for the user "ubuntu". After logging into the server it goes to ~/7daysded and executes the script restart.sh, so for the restart to work the server has to be set up accordingly. Alternatively you can of course put any other script there and name it restart.sh.